10.-Proyecto/PoCs/game/requirements.txt
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
cryptography==36.0.1
docutils==0.18.1
evdev==1.4.0
idna==3.3
importlib-metadata==4.10.0
jeepney==0.7.1
keyring==23.4.0
miguel-expansions==0.2
packaging==21.3
pkginfo==1.8.2
pycparser==2.21
Pygments==2.10.0
pynput==1.7.5
pyparsing==3.0.6
python-xlib==0.31
readme-renderer==32.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
tqdm==4.62.3
twine==3.7.1
urllib3==1.26.7
webencodings==0.5.1
zipp==3.6.0
